Update do_work.ps1
Don't fail if no files are present in folder.

+++ do_work.ps1
@@ -1,8 +1,10 @@
 $taglib = "d:\tools\taglib-sharp.dll"
 [system.reflection.assembly]::loadfile($taglib)
 
-$basepath = "h:\mp3"
+$basepath = "d:\soulseek-downloads\Known"
+$basepath = "d:\soulseek-downloads\Sorted"
 $basepath = "d:\mp3"
+$basepath = "h:\mp3"
 $basepath = "d:\soulseek-downloads\complete"
 
 #$a = Get-ChildItem d:\mp3 -recurse | Where-Object {$_.PSIsContainer -eq $True}
@@ -24,8 +26,14 @@ $basepath = "d:\soulseek-downloads\complete"
 #$a = Get-ChildItem "D:\soulseek-downloads\Pre-Sorted\Elephant Gym" -recurse | Where-Object {$_.PSIsContainer -eq $True}
 
 function cleanup {
-	$a = Get-ChildItem $basepath -recurse | Where-Object {$_.PSIsContainer -eq $True}
-	$a | Where-Object {$_.GetFiles().Count -eq 0 -and $_.GetDirectories().count -eq 0} | Remove-Item
+	$a = $args[0]
+	if (-not $a -eq $null) {
+		write-host "Cleaning up emtpy folders"
+		$a | Where-Object {$_.GetFiles().Count -eq 0 -and $_.GetDirectories().count -eq 0} | Remove-Item
+	}
+}
+function get_list {
+	write-host "`nGetting file list"
 	$a = Get-ChildItem $basepath -recurse | Where-Object {$_.PSIsContainer -eq $True}
 	return $a
 }
@@ -83,148 +91,162 @@ function do_work {
 	$total = 0
 	$inc = 0
 	$last = -1
-	$total = ($a | Where-Object {$_.GetFiles().Count -ne 0 -and $_.GetDirectories().count -eq 0} | get-childitem | Measure-Object).Count
-	write-host "`nProcessing $total items`n"
+	if (-not $a -eq $null) {	
+		$total = ($a | Where-Object {$_.GetFiles().Count -ne 0 -and $_.GetDirectories().count -eq 0} | get-childitem | Measure-Object).Count
+		write-host "`nProcessing $total items`n"
+		$a | Where-Object {$_.GetFiles().Count -ne 0 -and $_.GetDirectories().count -eq 0} | get-childitem | ForEach-Object {
+			$inc = $inc + 1;
+			$progress = [math]::floor($inc / $total * 100)
+			if ($progress -ne $last) {
+				write-host "`rProgress: $progress %" -nonewline
+				$last = $progress
+			}
 
-	$a | Where-Object {$_.GetFiles().Count -ne 0 -and $_.GetDirectories().count -eq 0} | get-childitem | ForEach-Object {
-		$inc = $inc + 1;
-		$progress = [math]::floor($inc / $total * 100)
-		if ($progress -ne $last) {
-			write-host "`rProgress: $progress %" -nonewline
-			$last = $progress
-		}
-
-		if ( $_.FullName -like '*mp3') {
-		} elseif ( $_.FullName -like '*mp2') {
-		} elseif ( $_.FullName -like '*cue' ) {
-			write-host "Split => " $_.FullName
-		} elseif ( $_.FullName -like '*wav' ) {
-			$d=$_.FullName -replace "wav$", "mp3"
-			write-host $_.FullName $d
-			if (-Not (Test-Path -LiteralPath $d)) {
-				& 'C:\Program Files (x86)\Lame\lame.exe' $_.FullName $d -b 320
-			}
-			copy_tag wav $_.FullName
-			Remove-Item -LiteralPath $_.FullName
-		} elseif ( $_.FullName -like '*aif') {
-			$d=$_.FullName -replace "aif$", "wav"
-			$e=$_.FullName -replace "aif$", "mp3"
-			if (-not (Test-Path -LiteralPath "$e")) {
-			  & "d:\tools\ffmpeg.exe" -i $_.FullName $d 
-			  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
-			  Remove-Item -LiteralPath $d
-			}
-			copy_tag aif $_.FullName
-			Remove-Item -LiteralPath $_.FullName
-		} elseif ( $_.FullName -like '*aiff') {
-			$d=$_.FullName -replace "aiff$", "wav"
-			$e=$_.FullName -replace "aiff$", "mp3"
-			if (-not (Test-Path -LiteralPath "$e")) {
-			  & "d:\tools\ffmpeg.exe" -i $_.FullName $d
-			  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
-			  Remove-Item -LiteralPath $d
-			}
-			copy_tag aiff $_.FullName
-			Remove-Item -LiteralPath $_.FullName
-		} elseif ( $_.FullName -like '*aac') {
-			$d=$_.FullName -replace "aac$", "wav"
-			$e=$_.FullName -replace "aac$", "mp3"
-			if (-not (Test-Path -LiteralPath "$e")) {
-			  & "d:\tools\ffmpeg.exe" -i $_.FullName $d 
-			  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
-			  Remove-Item -LiteralPath $d
-			}
-			copy_tag aac $_.FullName
-			Remove-Item -LiteralPath $_.FullName
-		} elseif ( $_.FullName -like '*m4a') {
-			$d=$_.FullName -replace "m4a$", "wav"
-			$e=$_.FullName -replace "m4a$", "mp3"
-			if (-not (Test-Path -LiteralPath "$e")) {
-			  & "d:\tools\ffmpeg.exe" -i $_.FullName $d 
-			  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
-			  Remove-Item -LiteralPath $d
-			}
-			copy_tag m4a $_.FullName
-			Remove-Item -LiteralPath $_.FullName
-		} elseif ( $_.FullName -like '*ogg') {
-			$d=$_.FullName -replace "ogg$", "wav"
-			$e=$_.FullName -replace "ogg$", "mp3"
-			if (-not (Test-Path -LiteralPath "$e")) {
-			  & "d:\tools\ffmpeg.exe" -i $_.FullName $d 
-			  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
-			  Remove-Item -LiteralPath $d
-			}
-			copy_tag ogg $_.FullName
-			Remove-Item -LiteralPath $_.FullName
-		} elseif ( $_.FullName -like '*ape') {
-			$d=$_.FullName -replace "ape$", "wav"
-			$e=$_.FullName -replace "ape$", "mp3"
-			if (-not (Test-Path -LiteralPath "$e")) {
-			  & "C:\Program Files (x86)\Monkey's Audio\MAC.exe" $_.FullName $d -d
-			  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
-			  Remove-Item -LiteralPath $d
-			}
-			copy_tag ape $_.FullName
-			Remove-Item -LiteralPath $_.FullName
-		} elseif ( $_.FullName -like '*flac' ) {
-			$d=$_.FullName -replace "flac$", "wav"
-			$e=$_.FullName -replace "flac$", "mp3"
-			if (-not (Test-Path -LiteralPath "$e")) {
-			  & 'd:\tools\flac.exe' -d $_.FullName -F
-			  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
-			  Remove-Item -LiteralPath $d
-			}
-			copy_tag flac $_.FullName
-			Remove-Item -LiteralPath $_.FullName
-		} elseif ( $_.FullName -like '*wma') {
-			$d=$_.FullName -replace "wma$", "wav"
-			$e=$_.FullName -replace "wma$", "mp3"
-			if (-not (Test-Path -LiteralPath "$e")) {
-			  & "d:\tools\ffmpeg.exe" -i $_.FullName $d 
-			  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
-			  Remove-Item -LiteralPath $d
-			}
-			copy_tag wma $_.FullName
-			Remove-Item -LiteralPath $_.FullName
-		} elseif ( $_.FullName -like '*wv' ) {
-			$d=$_.FullName -replace "wv$", "wav"
-			$e=$_.FullName -replace "wv$", "mp3"
-			if (-not (Test-Path -LiteralPath "$e")) {
-			  & 'd:\tools\wvunpack.exe'  --wav $_.FullName $d
-			  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
-			  Remove-Item -LiteralPath $d
-			}
-			copy_tag wv $_.FullName
-			Remove-Item -LiteralPath $_.FullName
-		} elseif ( $_.FullName -like '*rar' ) {
-			$7zoutput = Split-Path $_.FullName -Parent
-			#write-host $7zoutput
-			& 'C:\Program Files\7-Zip\7z.exe' e -yo"$7zoutput" $_.FullName 
-			Remove-Item -LiteralPath $_.FullName
-		} elseif ( $_.FullName -like '*zip' ) {
-			$7zoutput = Split-Path $_.FullName -Parent
-			#write-host $7zoutput
-			& 'C:\Program Files\7-Zip\7z.exe' e -yo"$7zoutput" $_.FullName 
-			Remove-Item -LiteralPath $_.FullName
-		} ElseIf ($_.FullName -like '*avi' -or $_.FullName -like '*mov' -or $_.FullName -like '*mpeg' -or $_.FullName -like '*mpg' -or $_.FullName -like '*wmv' -or $_.FullName -like '*flv') {
-		} else {
-			if ($_.FullName -like '*.original_epub' -or $_.FullName -like '*.prc' -or $_.FullName -like '*.azw3' -or $_.FullName -like '*.lit' -or $_.FullName -like '*.mobi' -or $_.FullName -like '*.opf' -or $_.FullName -like '*psd' -or $_.FullName -like '*dsn' -or $_.FullName -like '*acl' -or $_.FullName -like '*css' -or $_.FullName -like '*cld' -or $_.FullName -like '*sfk' -or $_.FullName -like '*onetoc2' -or $_.FullName -like '*csv' -or $_.FullName -like '*conf' -or $_.FullName -like '*ini'  -or $_.FullName -like '*lrc' -or $_.FullName -like '*bmp' -or $_.FullName -like '*ico' -or $_.FullName -like '*jpeg' -or $_.FullName -like '*tif' -or $_.FullName -like '*m3u' -or $_.FullName -like '*doc'  -or $_.FullName -like '*png' -or $_.FullName -like '*gif' -or $_.FullName -like '*jpg' -or $_.FullName -like '*DS_Store' -or $_.FullName -like '*nzb' -or $_.FullName -like '*torrent' -or $_.FullName -like '*rtf' -or $_.FullName -like '*db' -or $_.FullName -like '*html' -or $_.FullName -like '*log' -or $_.FullName -like '*nfo' -or $_.FullName -like '*docx' -or $_.FullName -like '*pdf' -or $_.FullName -like '*tqd' -or $_.FullName -like '*txt' -or $_.FullName -like '*md5' -or $_.FullName -like '*dat' -or $_.FullName -like '*asd' -or $_.FullName -like '*htm' -or $_.FullName -like '*yaml' -or $_.FullName -like '*hash' -or $_.FullName -like '*sfv' -or $_.FullName -like '*sls' -or $_.FullName -like '*url' -or $_.FullName -like '*lnk' -or $_.FullName -like '*xml' -or $_.FullName -like '*mht' -or $_.FullName -like '*directory' -or $_.FullName -like '*micro' -or $_.FullName -like '*bak' -or $_.FullName -like '*fpl' -or $_.FullName -like '*nra' -or $_.FullName -like '*diz' -or $_.FullName -like '*m3u8' -or $_.FullName -like '*mxm' -or $_.FullName -like '*b52' -or $_.FullName -like '*jsp' -or $_.FullName -like '*lfml' -or $_.FullName -like '*thm' -or $_.FullName -like '*webloc' -or $_.FullName -like '*accurip' -or $_.FullName -like '*pamp' -or $_.FullName -like '*m3u8'-or $_.FullName -like '*par2'-or $_.FullName -like '*webp' -or $_.FullName -like '*bat' -or $_.FullName -like '*tmp' -or $_.FullName -like '*cfg') {
-			  write-host "*$_.FullName"
-			  remove-item  -LiteralPath $_.FullName
+			if ( $_.FullName -like '*mp3') {
+			} elseif ( $_.FullName -like '*mp2') {
+			} elseif ( $_.FullName -like '*cue' ) {
+				write-host "Split => " $_.FullName
+			} elseif ( $_.FullName -like '*wav' ) {
+				$d=$_.FullName -replace "wav$", "mp3"
+				write-host $_.FullName $d
+				if (-Not (Test-Path -LiteralPath $d)) {
+					& 'C:\Program Files (x86)\Lame\lame.exe' $_.FullName $d -b 320
+				}
+				copy_tag wav $_.FullName
+				Remove-Item -LiteralPath $_.FullName
+			} elseif ( $_.FullName -like '*aif') {
+				$d=$_.FullName -replace "aif$", "wav"
+				$e=$_.FullName -replace "aif$", "mp3"
+				if (-not (Test-Path -LiteralPath "$e")) {
+				  & "d:\tools\ffmpeg.exe" -i $_.FullName $d 
+				  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
+				  Remove-Item -LiteralPath $d
+				}
+				copy_tag aif $_.FullName
+				Remove-Item -LiteralPath $_.FullName
+			} elseif ( $_.FullName -like '*aiff') {
+				$d=$_.FullName -replace "aiff$", "wav"
+				$e=$_.FullName -replace "aiff$", "mp3"
+				if (-not (Test-Path -LiteralPath "$e")) {
+				  & "d:\tools\ffmpeg.exe" -i $_.FullName $d
+				  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
+				  Remove-Item -LiteralPath $d
+				}
+				copy_tag aiff $_.FullName
+				Remove-Item -LiteralPath $_.FullName
+			} elseif ( $_.FullName -like '*aac') {
+				$d=$_.FullName -replace "aac$", "wav"
+				$e=$_.FullName -replace "aac$", "mp3"
+				if (-not (Test-Path -LiteralPath "$e")) {
+				  & "d:\tools\ffmpeg.exe" -i $_.FullName $d 
+				  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
+				  Remove-Item -LiteralPath $d
+				}
+				copy_tag aac $_.FullName
+				Remove-Item -LiteralPath $_.FullName
+			} elseif ( $_.FullName -like '*m4a') {
+				$d=$_.FullName -replace "m4a$", "wav"
+				$e=$_.FullName -replace "m4a$", "mp3"
+				if (-not (Test-Path -LiteralPath "$e")) {
+				  & "d:\tools\ffmpeg.exe" -i $_.FullName $d 
+				  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
+				  Remove-Item -LiteralPath $d
+				}
+				copy_tag m4a $_.FullName
+				Remove-Item -LiteralPath $_.FullName
+			} elseif ( $_.FullName -like '*ogg') {
+				$d=$_.FullName -replace "ogg$", "wav"
+				$e=$_.FullName -replace "ogg$", "mp3"
+				if (-not (Test-Path -LiteralPath "$e")) {
+				  & "d:\tools\ffmpeg.exe" -i $_.FullName $d 
+				  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
+				  Remove-Item -LiteralPath $d
+				}
+				copy_tag ogg $_.FullName
+				Remove-Item -LiteralPath $_.FullName
+			} elseif ( $_.FullName -like '*ape') {
+				$d=$_.FullName -replace "ape$", "wav"
+				$e=$_.FullName -replace "ape$", "mp3"
+				if (-not (Test-Path -LiteralPath "$e")) {
+				  & "C:\Program Files (x86)\Monkey's Audio\MAC.exe" $_.FullName $d -d
+				  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
+				  Remove-Item -LiteralPath $d
+				}
+				copy_tag ape $_.FullName
+				Remove-Item -LiteralPath $_.FullName
+			} elseif ( $_.FullName -like '*flac' ) {
+				$d=$_.FullName -replace "flac$", "wav"
+				$e=$_.FullName -replace "flac$", "mp3"
+				if (-not (Test-Path -LiteralPath "$e")) {
+				  & 'd:\tools\flac.exe' -d $_.FullName -F
+				  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
+				  Remove-Item -LiteralPath $d
+				}
+				copy_tag flac $_.FullName
+				Remove-Item -LiteralPath $_.FullName
+			} elseif ( $_.FullName -like '*wma') {
+				$d=$_.FullName -replace "wma$", "wav"
+				$e=$_.FullName -replace "wma$", "mp3"
+				if (-not (Test-Path -LiteralPath "$e")) {
+				  & "d:\tools\ffmpeg.exe" -i $_.FullName $d 
+				  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
+				  Remove-Item -LiteralPath $d
+				}
+				copy_tag wma $_.FullName
+				Remove-Item -LiteralPath $_.FullName
+			} elseif ( $_.FullName -like '*wv' ) {
+				$d=$_.FullName -replace "wv$", "wav"
+				$e=$_.FullName -replace "wv$", "mp3"
+				if (-not (Test-Path -LiteralPath "$e")) {
+				  & 'd:\tools\wvunpack.exe'  --wav $_.FullName $d
+				  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
+				  Remove-Item -LiteralPath $d
+				}
+				copy_tag wv $_.FullName
+				Remove-Item -LiteralPath $_.FullName
+			} elseif ( $_.FullName -like '*opus' ) {
+				$d=$_.FullName -replace "opus$", "wav"
+				$e=$_.FullName -replace "opus$", "mp3"
+				if (-not (Test-Path -LiteralPath "$e")) {
+				  & 'd:\tools\opusdec.exe' $_.FullName $d
+				  & 'C:\Program Files (x86)\Lame\lame.exe' $d $e -b 320
+				  Remove-Item -LiteralPath $d
+				}
+				copy_tag wv $_.FullName
+				Remove-Item -LiteralPath $_.FullName
+			} elseif ( $_.FullName -like '*rar' ) {
+				$7zoutput = Split-Path $_.FullName -Parent
+				#write-host $7zoutput
+				& 'C:\Program Files\7-Zip\7z.exe' e -yo"$7zoutput" $_.FullName 
+				Remove-Item -LiteralPath $_.FullName
+			} elseif ( $_.FullName -like '*zip' ) {
+				$7zoutput = Split-Path $_.FullName -Parent
+				#write-host $7zoutput
+				& 'C:\Program Files\7-Zip\7z.exe' e -yo"$7zoutput" $_.FullName 
+				Remove-Item -LiteralPath $_.FullName
+			} elseif ( $_.FullName -like '*7z' ) {
+				$7zoutput = Split-Path $_.FullName -Parent
+				#write-host $7zoutput
+				& 'C:\Program Files\7-Zip\7z.exe' e -yo"$7zoutput" $_.FullName 
+				Remove-Item -LiteralPath $_.FullName			
+			} ElseIf ($_.FullName -like '*avi' -or $_.FullName -like '*mp4' -or $_.FullName -like '*m4v' -or $_.FullName -like '*mov' -or $_.FullName -like '*mpeg' -or $_.FullName -like '*mpg' -or $_.FullName -like '*wmv' -or $_.FullName -like '*flv') {
 			} else {
-			  if (-not ((GET-ITEM $_.FullName) -is [system.io.directoryinfo])) {
-				write-host $_.FullName
-			  }
+				if ($_.FullName -like '*.srr' -or $_.FullName -like '*.jpe' -or $_.FullName -like '*.mht' -or $_.FullName -like '*.original_epub' -or $_.FullName -like '*.part' -or $_.FullName -like '*.cdr' -or $_.FullName -like '*.wps' -or $_.FullName -like '*.reapeaks' -or $_.FullName -like '*.ffp' -or $_.FullName -like '*.pls' -or $_.FullName -like '*.odt' -or $_.FullName -like '*.prc' -or $_.FullName -like '*.azw3' -or $_.FullName -like '*.lit' -or $_.FullName -like '*.mobi' -or $_.FullName -like '*.opf' -or $_.FullName -like '*psd' -or $_.FullName -like '*dsn' -or $_.FullName -like '*acl' -or $_.FullName -like '*css' -or $_.FullName -like '*cld' -or $_.FullName -like '*sfk' -or $_.FullName -like '*onetoc2' -or $_.FullName -like '*csv' -or $_.FullName -like '*conf' -or $_.FullName -like '*ini'  -or $_.FullName -like '*lrc' -or $_.FullName -like '*bmp' -or $_.FullName -like '*ico' -or $_.FullName -like '*jpeg' -or $_.FullName -like '*tif' -or $_.FullName -like '*m3u' -or $_.FullName -like '*doc'  -or $_.FullName -like '*png' -or $_.FullName -like '*gif' -or $_.FullName -like '*jpg' -or $_.FullName -like '*DS_Store' -or $_.FullName -like '*nzb' -or $_.FullName -like '*torrent' -or $_.FullName -like '*rtf' -or $_.FullName -like '*db' -or $_.FullName -like '*html' -or $_.FullName -like '*log' -or $_.FullName -like '*nfo' -or $_.FullName -like '*docx' -or $_.FullName -like '*pdf' -or $_.FullName -like '*tqd' -or $_.FullName -like '*txt' -or $_.FullName -like '*md5' -or $_.FullName -like '*dat' -or $_.FullName -like '*asd' -or $_.FullName -like '*htm' -or $_.FullName -like '*yaml' -or $_.FullName -like '*hash' -or $_.FullName -like '*sfv' -or $_.FullName -like '*sls' -or $_.FullName -like '*url' -or $_.FullName -like '*lnk' -or $_.FullName -like '*xml' -or $_.FullName -like '*mht' -or $_.FullName -like '*directory' -or $_.FullName -like '*micro' -or $_.FullName -like '*bak' -or $_.FullName -like '*fpl' -or $_.FullName -like '*nra' -or $_.FullName -like '*diz' -or $_.FullName -like '*m3u8' -or $_.FullName -like '*mxm' -or $_.FullName -like '*b52' -or $_.FullName -like '*jsp' -or $_.FullName -like '*lfml' -or $_.FullName -like '*thm' -or $_.FullName -like '*webloc' -or $_.FullName -like '*accurip' -or $_.FullName -like '*pamp' -or $_.FullName -like '*m3u8'-or $_.FullName -like '*par2'-or $_.FullName -like '*webp' -or $_.FullName -like '*bat' -or $_.FullName -like '*tmp' -or $_.FullName -like '*cfg') {
+				  write-host "Deleting $_"
+				  remove-item  -LiteralPath $_.FullName
+				} else {
+				  if (-not ((GET-ITEM $_.FullName) -is [system.io.directoryinfo])) {
+					write-host $_.FullName
+				  }
+				}
 			}
 		}
+	} else {
+		write-host "Nothing to process."
 	}
 }
 
-$a = cleanup
+$a = get_list
 do_work $a
-$a = cleanup
-
-$basepath = "d:\soulseek-downloads\Known"
-$a = cleanup
-
-                              
+$a = get_list
+cleanup $a
